selenium updated for jenkins
